GUM-32: updated navigation structure
