+++ src/common/enums/navigation/app-route.enum.ts
@@ -5,7 +5,10 @@ enum AppRoute {
   OKR = 'OKR',
   PROFILE = 'Profile',
   APP = 'App',
+  APP_TUBS = 'AppTubs',
   ONBOARDING_SETUP = 'OnboardingSetup',
+  ADD_OKR = 'AddOKR',
+  ADD_KEY_RESULT = 'AddKeyResult',
 }
 
 export { AppRoute };
