+++ src/common/enums/navigation/index.ts
@@ -2,4 +2,3 @@ export { AuthRoute } from './auth-route.enum';
 export { AppRoute } from './app-route.enum';
 export { ProfileRoute } from './profile-route.enum';
 export { OnboardingRoute } from './onboarding-route.enum';
-export { OKRRoute } from './okr-route.enum';
