--- src/common/enums/navigation/okr-route.enum.ts
@@ -1,7 +1,0 @@
-enum OKRRoute {
-  OKR = 'OKRScreen',
-  ADD_OKR = 'AddOKR',
-  ADD_KEY_RESULT = 'AddKeyResult',
-}
-
-export { OKRRoute };
