+++ src/common/types/navigation/app-tabs-param.ts
@@ -1,0 +1,11 @@
+import { AppRoute } from 'src/common/enums';
+
+type AppTabsParamList = {
+  [AppRoute.HOME]: undefined;
+  [AppRoute.EXPLORE]: undefined;
+  [AppRoute.OPPORTUNITIES]: undefined;
+  [AppRoute.OKR]: undefined;
+  [AppRoute.PROFILE]: undefined;
+};
+
+export type { AppTabsParamList };
