+++ src/common/types/navigation/index.ts
@@ -1,2 +1,4 @@
 export * from './onboarding-stack-param';
-export * from './okr-stack-param';
+export * from './app-stack-param';
+export * from './app-tabs-param';
+export * from './root-stack-param';
