--- src/common/types/navigation/okr-stack-param.ts
@@ -1,11 +1,0 @@
-import { OKRRoute } from 'src/common/enums';
-
-type OKRStackParamList = {
-  [OKRRoute.OKR]: undefined;
-  [OKRRoute.ADD_OKR]: {
-    isTeamOkr: boolean;
-  };
-  [OKRRoute.ADD_KEY_RESULT]: undefined;
-};
-
-export type { OKRStackParamList };
