+++ src/navigation/app-tabs/index.ts
@@ -1,0 +1,1 @@
+export { default as AppTabsNavigation } from './app-tabs-navigation';
