+++ src/navigation/app/index.ts
@@ -1,0 +1,1 @@
+export { default as AppNavigation } from './app-navigation';
