--- src/navigation/okr/index.ts
@@ -1,1 +1,0 @@
-export { default as OKRNavigation } from './okr-navigation';
